Address review comments from March 15
