+++ api/src/test/java/org/projectnessie/tools/catalog/migration/api/ITNessieToHiveCatalogMigrator.java
@@ -13,19 +13,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.projectnessie.tools.catalog.migration.api.test;
+package org.projectnessie.tools.catalog.migration.api;
 
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Disabled;
+import org.projectnessie.tools.catalog.migration.api.test.HiveMetaStoreRunner;
 
-public class ITHadoopToHiveCatalogMigrator extends AbstractTestCatalogMigrator {
+public class ITNessieToHiveCatalogMigrator extends AbstractTestCatalogMigrator {
+
+  protected static final int NESSIE_PORT = Integer.getInteger("quarkus.http.test-port", 19121);
+
+  protected static String nessieUri = String.format("http://localhost:%d/api/v1", NESSIE_PORT);
 
   @BeforeAll
   protected static void setup() throws Exception {
     HiveMetaStoreRunner.startMetastore();
 
-    catalog1 = createHadoopCatalog(warehouse2.toAbsolutePath().toString(), "hadoop");
+    catalog1 = createNessieCatalog(warehouse2.toAbsolutePath().toString(), nessieUri);
     catalog2 = HiveMetaStoreRunner.hiveCatalog();
 
     createNamespaces();
