+++ cli/src/main/resources/logback.xml
@@ -16,7 +16,8 @@
     limitations under the License.
 
 -->
-<configuration debug="true">
+<configuration>
+  <statusListener class="ch.qos.logback.core.status.NopStatusListener"/>
   <property name="catalog.migration.log.dir" value="${catalog.migration.log.dir:-logs}"/>
 
   <appender name="file" class="ch.qos.logback.core.FileAppender">
@@ -28,6 +29,7 @@
   </appender>
 
   <appender name="console" class="ch.qos.logback.core.ConsoleAppender">
+    <withJansi>true</withJansi>
     <encoder>
       <pattern>%highlight(%-5level) - %msg%n</pattern>
     </encoder>
